[hw17-multiprocess] - replaced CountDownLatch with synchronized block and changed blockingStub to stub
